+++ models/donation.go
@@ -60,6 +60,7 @@ type PayByPrimeDonation struct {
 	Status      string     `gorm:"type:ENUM('paying','paid','fail');not null" json:"status"`
 	UpdatedAt   time.Time  `json:"updated_at"`
 	UserID      uint       `gorm:"type:int(10);unsigned;not null" json:"user_id"`
+	IsAnonymous null.Bool  `gorm:"type:tinyint(1);default:0" json:"is_anonymous"`
 }
 
 type PayByCardTokenDonation struct {
@@ -120,4 +121,5 @@ type PeriodicDonation struct {
 	ToFeedback    null.Bool  `gorm:"type:tinyint(1);default:1" json:"to_feedback"`
 	UpdatedAt     time.Time  `json:"updated_at"`
 	UserID        uint       `gorm:"type:int(10) unsigned;not null" json:"user_id"`
+	IsAnonymous   null.Bool  `gorm:"type:tinyint(1);default:0" json:"is_anonymous"`
 }